% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{delve_query}
\alias{delve_query}
\title{Query DELVE API}
\usage{
delve_query(
  project_id,
  dataset_id,
  version,
  filename,
  user_token = get_user_token()
)
}
\arguments{
\item{project_id}{ProjectID (integer)}

\item{dataset_id}{DataSetID (integer)}

\item{version}{Version (integer)}

\item{filename}{Filename}

\item{user_token}{DELVE user token}
}
\description{
Submit GET request to DELVE API
}
\details{
Returns response to DELVE query as a data frame
}
