% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_datasets}
\alias{get_datasets}
\title{Get datasets}
\usage{
get_datasets(user_token = get_user_token(), use_qa = FALSE)
}
\arguments{
\item{user_token}{DELVE user token}

\item{use_qa}{Boolean flag to use QA version of DELVE}
}
\description{
Get DELVE dataset metadata as a dataframe
}
