% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_filetypes}
\alias{get_filetypes}
\title{Get file types}
\usage{
get_filetypes(user_token = get_user_token(), use_qa = FALSE)
}
\arguments{
\item{user_token}{DELVE user token}

\item{use_qa}{Boolean flag to use QA version of DELVE}
}
\description{
Get DELVE filetype metadata as a dataframe
}
