% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_metadata.R
\name{get_projects}
\alias{get_projects}
\alias{get_datasets}
\alias{get_dataset_types}
\alias{get_file_types}
\alias{get_keywords}
\title{Get metadata}
\usage{
get_projects(user_token = get_user_token(), use_qa = FALSE)

get_datasets(user_token = get_user_token(), use_qa = FALSE)

get_dataset_types(user_token = get_user_token(), use_qa = FALSE)

get_file_types(user_token = get_user_token(), use_qa = FALSE)

get_keywords(user_token = get_user_token(), use_qa = FALSE)
}
\arguments{
\item{user_token}{DELVE user token}

\item{use_qa}{Boolean flag to use QA version of DELVE}
}
\description{
Get DELVE metadata for projects, datasets, dataset types, file types, and keywords.
}
