% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/token.R
\name{set_user_token}
\alias{set_user_token}
\alias{get_user_token}
\title{Set or get DELVE API token}
\usage{
set_user_token(token = NULL)

get_user_token()
}
\description{
Set or get DELVE API token as environmental variable
}
