% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_dataset_file}
\alias{get_dataset_file}
\title{Get dataset file}
\usage{
get_dataset_file(
  project_id,
  dataset_id,
  version,
  filename,
  user_token = get_user_token()
)
}
\arguments{
\item{project_id}{ProjectID (integer)}

\item{dataset_id}{DataSetID (integer)}

\item{version}{Version (integer)}

\item{filename}{Filename}

\item{user_token}{DELVE user token}
}
\description{
Get a specific version of a file from within a DELVE project and dataset
}
