% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_projects}
\alias{get_projects}
\title{Get projects}
\usage{
get_projects(user_token = get_user_token(), use_qa = FALSE)
}
\arguments{
\item{user_token}{DELVE user token}

\item{use_qa}{Boolean flag to use QA version of DELVE}
}
\description{
Get list of DELVE projects as a dataframe
}
